backend/requirements.txt
fastapi
uvicorn[standard]
google-cloud-aiplatform
googlemaps
python-dotenv
pydantic
requests
slowapi
